MAINT: Move report code (#9749)
* MAINT: Move report code

* FIX: No need for import private
