+++ mne/report/__init__.py
@@ -1,0 +1,1 @@
+from .report import Report, open_report, _ReportScraper
